[SelectionPtr]Cache the ListSelectionModel for the accessor
TabStripModel::selection_model() returns a ListSelectionModel object but
it is internally using a SelectionState, and so we need to convert this
object to a ListSelectionModel. This is linear in the number of tabs, so
it is concerning if there are clients who spam calls to
TabStripModel::selection_model(). As a precautionary measure we can
cache a ListSelectionModel object and generate it only when necessary
(when the indices of the selection are potentially updated).

There are no tests in this file since after looking at
tab_strip_model_unittest.cc, these changes are already tested by the
ones there.

Change-Id: I58e2beb7eca900f6e1d1bf19b45ffcf02ed4823a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7281804
Commit-Queue: Dominic Austria <dominicaustria@google.com>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1563457}