Add log statements to debug flaky test case
testZeroPrefixSuggestions_Incognito recently started flaking in tablet
bots as a result of sometimes failing to open a new Incognito window.
This could be due to the new short-term fix of caching the screen size
in order to determine if the device is a tablet. Add 3 log statements to
see if the flakes are from reading an incorrect form factor.

Bug: 470234934
Change-Id: Ic6b0335b7afd94dece0ef53ae475fc19caa31452
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7355241
Owners-Override: Henrique Nakashima <hnakashima@chromium.org>
Commit-Queue: Michael Wu <mklw@google.com>
Reviewed-by: Henrique Nakashima <hnakashima@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1563480}