[ntp-simplification] Add UMA logging for NTP modules auto-removal
This CL adds UMA logging to track the number of times NTP modules are
automatically removed due to inactivity and when this action is undone.
Unit tests are added for coverage.

It introduces two histograms, `NewTabPage.Modules.AutoRemovedModuleId`
and `NewTabPage.Modules.AutoRemovalUndoneModuleId`, which record the
auto-removal and undone count of individual module with the persistent
hash of the module ID to align with UMA best practices.

Bug: 469839722, 469839616
Change-Id: I6c962a4a1ecb4c50292b1e1733c2b0c94d31d1ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7318743
Reviewed-by: Roman Arora <romanarora@chromium.org>
Commit-Queue: Bofeng Chen <bofengchen@google.com>
Cr-Commit-Position: refs/heads/main@{#1563517}