[Extensions] Use GetSenderWebContents instead of GetVisibleWebContents
ExtensionFunctionDispatcher::GetVisibleWebContents() is *only* called
from the Downloads API when we're about to show a dialog. However, this
has problems:
* In practice, GetVisibleWebContents() defaults in
  GetAssociatedWebContents in every case *except* for an
  ExtensionViewHost.
* GetAssociatedWebContents() returns the web contents for the delegate.
  While this has different implementations, it should always correspond
  to the WebContents associated with the extension function call (if
  any).
* In the case of an ExtensionViewHost, this defaults to
  GetAssociatedWebContents() for extension popups, and otherwise returns
  null.
* The only other time in which we use an ExtensionViewHost is for
  side panel extension views.

Thus, the practical effect of this method is to only return null for
side panel extension views. This has no bearing on whether a
WebContents is visible (since for e.g. background pages, the web
contents of the background page will still be returned, even though it's
not visible).

Additionally, this isn't necessarily deterministic. Both
ExtensionViewHost and ExtensionWebContentsObserver try to be the
delegate of the ExtensionFunctionDispatcher. If the
ExtensionWebContentsObserver "wins", the associated web contents will
be returned, even for side panel views.

Update the downloads API to instead use
ExtensionFunction::GetSenderWebContents(), which is the canonical way
for an ExtensionFunction to retrieve the web contents for the function
call. In practice, this has no behavioral change *except* for the
possible scenario of the side panel case described above (and then, only
if the ExtensionViewHost "won" as the delegate). The end result is that
there's a possibility more downloads API methods would succeed when
called from a side panel context -- which is a desirable fix.

Bug: 470896011
Change-Id: I57084ad5fc3a63cfabdf627b6b84d14545090787
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7362586
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1563542}