[Android][Signin] Add API for activityless sign-in flow
This change introduces an API in SigninAndHistorySyncActivityLauncher to
launch the bottom-sheet sign-in flow without a dedicated Activity.

It decouples BottomSheetSigninAndHistorySyncCoordinator from the current
sign-in Activity by:

* Splitting the Delegate interface into 2 parts
  * ActivityDelegate interface for actions handled by the sign-in
activity in the old flow (e.g. start add account activity)
  * Delegate interface for more generic actions in the old and the
new flow (like flow result handling)

* Using ActivityResultTracker to handle the "add account" activity
result in a new code path:
  * BottomSheetSigninAndHistorySyncCoordinator can now be created
without a config object, and will register the activity callback early
and skip the UI flow start in this case - This will be called by the
entry point.
  * A `start` methods accepting a config object is added to the same
coordinator. It starts the UI flow, and will be called when the sign-in
flow is triggered at the entry point level.

Test coverage will be added in a follow-up CL.

Bug: 437039516
Change-Id: If3d9ce126dde77d104e8910893c28edf8c123421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7302285
Commit-Queue: Menghan Yang <myuu@google.com>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1563651}